requirements.txt
flask==3.1.2 
mysql-connector-python==9.4.0 
flask-cors==6.0.1 
